arch/x86/include/arch/fit_table.h: add
Extract the definition from src/cpu/intel/fit/fit_table.c, change first
field to a union and add defines for types.  This provides a basis for
parsing FIT data from coreboot.

Upstream-Status: Pending
